+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - updated Quoizel sale on sale page
 - Home page sale banners
 - Sale page banners
+- Quoizel single product page coupon and logic
 
 ## [ released ] - 11.19.2020